+++ src/session/controller_itp.ml
@@ -117,7 +117,7 @@ let load_drivers c =
   let provers = Whyconf.get_provers config in
   let main = Whyconf.get_main config in
   Whyconf.Mprover.iter (fun _ p ->
-      let d = Driver.load_driver main env p in
+      let d = Driver.load_driver_for_prover main env p in
       Whyconf.Hprover.add c.controller_provers p.Whyconf.prover (p,d))
     provers