+++ CHANGES.md
@@ -17,13 +17,15 @@ API
   * source locations are now represented by both a start line number
     and an end line number. The fourth number is thus now the column
     number on the end line. :x:
-  * function load_driver moved from Whyconf to Driver :x:
   * several functions from Call_prover and Driver now take as input a
     `Whyconf.main` configuration type instead of directly take libdir
     and datadir. See the updated API examples for details. :x:
+  * Use of [Whyconf.load_driver] must be replaced by
+    [Driver.load_driver_for_prover] :x:
 
 MLCFG input language
   * attributes on the body of an MLCFG function are now supported
+
 
 Version 1.5.0, April 29, 2022
 -----------------------------